+++ frontend/src/hooks/api/userSecrets/mutations.ts
@@ -9,7 +9,7 @@ export const useCreateUserSecret = () => {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: async (inputData: TCreateUserSecretRequest) => {
-      const { data } = await apiRequest.post<TUserSecret>("/api/v1/secret-sharing", inputData);
+      const { data } = await apiRequest.post<TUserSecret>("/api/v1/user-secret", inputData);
       return data;
     },
     onSuccess: () => queryClient.invalidateQueries(userSecretKeys.allUserSecrets())
@@ -20,9 +20,7 @@ export const useDeleteUserSecret = () => {
   const queryClient = useQueryClient();
   return useMutation<TUserSecret, { message: string }, { userSecretId: string }>({
     mutationFn: async ({ userSecretId }: TDeleteUserSecretRequest) => {
-      const { data } = await apiRequest.delete<TUserSecret>(
-        `/api/v1/secret-sharing/${userSecretId}`
-      );
+      const { data } = await apiRequest.delete<TUserSecret>(`/api/v1/user-secret/${userSecretId}`);
       return data;
     },
     onSuccess: () => queryClient.invalidateQueries(userSecretKeys.allUserSecrets())
